+++ internal/gophermart/models/account.go
@@ -1,6 +1,6 @@
 package models
 
 type Account struct {
-	Balance   int `json:"current"`
-	Withdraws int `json:"withdrawn,omitempty"`
+	Balance   float64 `json:"current"`
+	Withdraws float64 `json:"withdrawn,omitempty"`
 }
